+++ HISTORY.rst
@@ -6,6 +6,14 @@ Important changes are emphasized.
 master
 ^^^^^^
 
+0.13.0
+^^^^^^
+
+- Support for recent versions of the ``fish`` shell!
+- Added ``--admin`` flag to ``install``, ``uninstall``, and ``update`` commands
+  to indicate elevated privileges have already been given.
+- Basic support for ``csh``/``tcsh`` shells.
+
 0.12.0
 ^^^^^^